refactor(learning): Stage3를 learning 도메인으로 통합
- Stage3 router/service/schema를 learning 도메인 내부로 이동
- /learning/stage3/* 외부 API 경로는 유지
- LearningRecordRepository를 Protocol과 Mongo 구현으로 분리
- 기존 learning.repository import는 호환용 re-export로 유지
- 도메인 구조 테스트와 모델 import 테스트 갱신
