+++ app/main.py
@@ -6,7 +6,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from app.domains.chat.router import router as chat_router
 from app.domains.system.router import router as system_router
 from app.domains.learning.content_router import router as learning_router
-from app.domains.stage3.router import router as stage3_router
+from app.domains.learning.stage3_router import router as stage3_router
 from app.domains.auth.router import router as auth_router
 from app.domains.agent.router import router as agent_router
 from app.domains.learning.router import router as learning_records_router
